refactor: move user authentication to services layer
- Move user auth functions from main.py to services/auth/user_auth.py

- Move auth utilities to services/auth/utils.py

- Update imports across affected modules
